Basic working library
Fixed a lot of strange things in the original library, such as the need
to pass in the length and lack of assert (may have been due to
limitations in older OpenSCAD), and removed the dependence on global $fa
value. Also renamed the internal modules with an underscore prefix to
make it a little more obvious which module you should be using.

Tested with and without base plate and it seems to work. Have not
printed.

The dots look kind of small to me, and the text is oriented along the Y
axis which seems weird. But