Default to upright braille, with lie_flat option
If you're just printing a label to stick on something else, it's better
ot print it upright so that the Braille will come out smoother. I
changed the default to reflect this, but added a lie_flat=true option
you can set to make the label lie flat (e.g. union with an existing
model).