Fix dot size + spacing
So in the original program the incorrect arg `radius` was used for
sphere() rather than the correct `r`, which meant the radius was being
used, and the library only worked right because of this since the set
radius was too small.

I've fixed this. I also took the opportunity to align the spacing to UEB
standards and to tweak the positioning in the plate a bit.

Printing this yielded readable Braille in both orientations although the
top was a bit flat on the lie_flat version.